+++ src/profile/games/cclcc/savemenu.cpp
@@ -39,7 +39,7 @@ void Configure() {
   SaveEntryPrimaryColor = EnsureGetMember<uint32_t>("SaveEntryPrimaryColor");
   SaveEntrySecondaryColor =
       EnsureGetMember<uint32_t>("SaveEntrySecondaryColor");
-  for (int i = 0; i < sizeof(SaveMenuTypeNames) / sizeof(*SaveMenuTypeNames);
+  for (size_t i = 0; i < sizeof(SaveMenuTypeNames) / sizeof(*SaveMenuTypeNames);
        i++) {
     UI::SaveMenuPageType menuType =
         UI::SaveMenuPageType::_from_integral_unchecked(i);
